requirements.txt
Flask==0.10.1
IpCamPy==0.3.0
Jinja2==2.9.6
MarkupSafe==1.0
Pillow==4.1.1
RPi.GPIO==0.6.3
Werkzeug==0.12.1
argparse==1.2.1
arrow==0.10.0
bottle==0.12.13
cffi==1.10.0
click==5.1
colorama==0.3.7
itsdangerous==0.24
lockfile==0.12.2
olefile==0.44
pil-compat==1.0.0
platformio==3.3.0
pycparser==2.17
pyserial==3.3
python-dateutil==2.6.0
requests==2.13.0
semantic-version==2.6.0
six==1.10.0
smbus-cffi==0.5.1
spidev==3.2
wsgiref==0.1.2
-e git+https://github.com/adafruit/Adafruit_Python_GPIO.git#egg=Adafruit-GPIO
-e git+https://github.com/adafruit/Adafruit_Python_PureIO.git#egg=Adafruit-PureIO
-e git+https://github.com/adafruit/Adafruit_Python_SSD1306.git#egg=Adafruit-SSD1306
